Add files via upload
Debug work.
Test part done. Gain an 85% accuracy on mnist.
This accuracy seems bad, but remember our model is based on a decision tree, who is again based on a single layer of fully connected layer and a sigmoid layer.

To do: I will try to use more complicated models, hoping the tree would self-balance. If the tree refuse to balance itself, I will do it manually, not by swapping, but by assigning classes to nodes.