Add files via upload
Major work in this version:

Median(instead of mean) of expectation among all classes in a node, judges where a node should go left or right. This helps balance the tree.

Suspending(delete, but can be recovered when this class arrives at this node again) unused classes recursively down the tree(instead of swap), is used to reduce height of tree.

To do:
try some 1D convolution network.